split DocumentEditor/relationship into shared/ui
