+++ packages/fields-document/src/DocumentEditor/insert-menu.tsx
@@ -21,7 +21,8 @@ import scrollIntoView from 'scroll-into-view-if-needed'
 import { ComponentBlockContext, insertComponentBlock } from './component-blocks'
 import { type ComponentBlock } from './component-blocks/api-shared'
 import { InlineDialog, ToolbarButton } from './primitives'
-import { type Relationships, useDocumentFieldRelationships } from './relationship'
+import { type Relationships } from './relationship-shared'
+import { useDocumentFieldRelationships } from './relationship'
 import { useToolbarState } from './toolbar-state'
 import { type ToolbarState, } from './toolbar-state-shared'
 import { insertNodesButReplaceIfSelectionIsAtEmptyParagraphOrHeading } from './utils'
