+++ packages/fields-document/src/relationship-data.tsx
@@ -10,7 +10,7 @@ import {
 } from './DocumentEditor/component-blocks/api-shared'
 
 import { assertNever } from './DocumentEditor/component-blocks/utils'
-import { type Relationships } from './DocumentEditor/relationship'
+import { type Relationships } from './DocumentEditor/relationship-shared'
 
 const labelFieldAlias = '____document_field_relationship_item_label'
 const idFieldAlias = '____document_field_relationship_item_id'
